# swagger/swagger.yaml
swagger: '2.0'
info:
  title: ShoppingApp
  description: Neosoft Shopping App
  version: 1.0.0
host: ''
schemes:
  - http
paths:
...
  /api/users/{userId}:
    get:
      tags:
        - User
      summary: Get user by Id
      description: Get user by Id
      operationId: getUserById
      produces:
        - application/json
      parameters:
      - name: "userId"
        in: "path"
        description: "_id of category to return"
        required: true
        type: "string"
      responses:
        '200':
          description: OK
          schema:
            type: string
      security: []
  /api/users:
    get:
      tags:
        - User
      summary: Users
      description: Users
      operationId: users
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            type: string
      security:
      - authorization: []
  /api/auth/login:
    post:
      tags:
        - User
      summary: User signin
      description: User signin
      operationId: userSignin
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          description: the "User" model
          required: true
          schema:
            $ref: '#/definitions/User'
      responses:
        '200':
          description: OK
          schema:
            type: string
      security: []
  /api/auth/sign-up:
    post:
      tags:
        - User
      summary: User signup
      description: User signup
      operationId: userSignup
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          description: the "User" model
          required: true
          schema:
            $ref: '#/definitions/User'
      responses:
        '200':
          description: OK
          schema:
            type: string
      security: []
...
definitions:
...
  User:
    description: Model for User
    required:
      - confirm_password
    properties:
      confirm_password:
        type: string
        x-example: jitendra123
      email:
        type: string
        x-example: jitendra.kumar@neosofttech.com
      first_name:
        type: string
        x-example: Jitendra
      gender:
        type: string
        x-example: Male
      last_name:
        type: string
        x-example: Kumar
      password:
        type: string
        x-example: jitendra123
...
securityDefinitions: 
  authorization:
    type: "apiKey"
    name: "authorization"
    in: "header"
security: []
...